tu puedes laura <3

+++ level_4/ft_itoa.c
@@ -3,7 +3,7 @@
 
 char	*ft_itoa(int nbr)
 {
-	(long)int n = nbr;
+	long n = (long)nbr;
       	int len = 0;
 	
 	if(n <= 0)
